Add AbstractPowerUserBlock
