+++ src/main/java/com/teamofelectrorealism/electrorealism/block/generator/AbstractGeneratorBlock.java
@@ -1,4 +1,5 @@
 package com.teamofelectrorealism.electrorealism.block.generator;
+import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.BaseEntityBlock;
 import net.minecraft.world.level.block.RenderShape;
@@ -8,8 +9,8 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 
-public abstract class GeneratorBlock extends BaseEntityBlock {
-    protected GeneratorBlock(Properties properties) {
+public abstract class AbstractGeneratorBlock extends BaseEntityBlock {
+    protected AbstractGeneratorBlock(Properties properties) {
         super(properties);
     }
 
@@ -20,4 +21,7 @@ public abstract class GeneratorBlock extends BaseEntityBlock {
 
     @Override
     public abstract @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType);
+
+
+    protected abstract void tick(Level level1, BlockPos pos, BlockState state1, AbstractGeneratorBlockEntity blockEntity);
 }
