+++ src/main/java/com/teamofelectrorealism/electrorealism/block/generator/solarpanel/SolarPanelBlock.java
@@ -2,7 +2,8 @@ package com.teamofelectrorealism.electrorealism.block.generator.solarpanel;
 
 import com.mojang.serialization.MapCodec;
 import com.teamofelectrorealism.electrorealism.block.ModBlockEntityTypes;
-import com.teamofelectrorealism.electrorealism.block.generator.GeneratorBlock;
+import com.teamofelectrorealism.electrorealism.block.generator.AbstractGeneratorBlock;
+import com.teamofelectrorealism.electrorealism.block.generator.AbstractGeneratorBlockEntity;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.BaseEntityBlock;
@@ -12,13 +13,14 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 
-public class SolarPanelBlock extends GeneratorBlock {
+public class SolarPanelBlock extends AbstractGeneratorBlock {
     public static final MapCodec<SolarPanelBlock> CODEC = simpleCodec(SolarPanelBlock::new);
     public SolarPanelBlock(Properties properties) {
         super(properties);
     }
 
-    private static void tick(Level level1, BlockPos pos, BlockState state1, SolarPanelBlockEntity blockEntity) {
+    @Override
+    protected void tick(Level level1, BlockPos pos, BlockState state1, AbstractGeneratorBlockEntity blockEntity) {
         blockEntity.tick(level1, pos, state1);
     }
 
@@ -32,7 +34,7 @@ public class SolarPanelBlock extends GeneratorBlock {
             return null;
         }
 
-        return createTickerHelper(blockEntityType, ModBlockEntityTypes.SOLAR_PANEL_BE.get(), SolarPanelBlock::tick);
+        return createTickerHelper(blockEntityType, ModBlockEntityTypes.SOLAR_PANEL_BE.get(), this::tick);
     }
 
     @Override
