+++ src/main/java/com/teamofelectrorealism/electrorealism/block/generator/solarpanel/SolarPanelBlockEntity.java
@@ -2,14 +2,14 @@ package com.teamofelectrorealism.electrorealism.block.generator.solarpanel;
 
 import com.teamofelectrorealism.electrorealism.block.IPowerReceiver;
 import com.teamofelectrorealism.electrorealism.block.ModBlockEntityTypes;
-import com.teamofelectrorealism.electrorealism.block.generator.GeneratorBlockEntity;
+import com.teamofelectrorealism.electrorealism.block.generator.AbstractGeneratorBlockEntity;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 
-public class SolarPanelBlockEntity extends GeneratorBlockEntity {
+public class SolarPanelBlockEntity extends AbstractGeneratorBlockEntity {
     private final int voltage = 230;
 
     public SolarPanelBlockEntity(BlockPos pos, BlockState blockState) {
