+++ src/main/java/com/teamofelectrorealism/electrorealism/block/user/AbstractPowerUserBlock.java
@@ -1,0 +1,27 @@
+package com.teamofelectrorealism.electrorealism.block.user;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.BaseEntityBlock;
+import net.minecraft.world.level.block.RenderShape;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityTicker;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.Nullable;
+
+public abstract class AbstractPowerUserBlock extends BaseEntityBlock {
+    protected AbstractPowerUserBlock(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    public RenderShape getRenderShape(BlockState state) {
+        return RenderShape.MODEL;
+    }
+
+    @Override
+    public abstract @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType);
+
+    protected abstract void tick(Level level1, BlockPos pos, BlockState state1, AbstractPowerUserBlockEntity blockEntity);
+}
